requirements.txt
#matplotlib
#numpy
#pandas
openai>=1.0.0,<2.0.0
python-dotenv
Flask==3.0.3
requests
chess
python-chess
Pillow